requirements.txt
python-telegram-bot==20.4
google-cloud-api-keys==0.5.3
google-cloud-dialogflow==2.23.3